+++ inout/src/reserved.rs
@@ -4,7 +4,7 @@ use core::{marker::PhantomData, slice};
 #[cfg(feature = "block-padding")]
 use {
     crate::{InOut, errors::PadError},
-    block_padding::{PadType, Padding},
+    block_padding::Padding,
     hybrid_array::{Array, ArraySize},
 };
 
@@ -160,55 +160,49 @@ impl<'inp, 'out> InOutBufReserved<'inp, 'out, u8> {
     {
         let bs = BS::USIZE;
         let blocks_len = self.in_len / bs;
-        let tail_len = self.in_len - bs * blocks_len;
+        let (blocks, tail_block) = P::pad_detached(self.get_in()).map_err(|_| PadError)?;
+
+        assert_eq!(blocks.len(), blocks_len);
+
+        let out_len = self.out_len;
+        let (in_ptr, out_ptr) = self.into_raw();
+
         let blocks = unsafe {
             InOutBuf::from_raw(
-                self.in_ptr as *const Array<u8, BS>,
-                self.out_ptr as *mut Array<u8, BS>,
+                in_ptr.cast::<Array<u8, BS>>(),
+                out_ptr.cast::<Array<u8, BS>>(),
                 blocks_len,
             )
         };
-        let mut tail_in = Array::<u8, BS>::default();
-        let tail_out = match P::TYPE {
-            PadType::NoPadding | PadType::Ambiguous if tail_len == 0 => None,
-            PadType::NoPadding => return Err(PadError),
-            PadType::Reversible | PadType::Ambiguous => {
-                let blen = bs * blocks_len;
-                let res_len = blen + bs;
-                if res_len > self.out_len {
-                    return Err(PadError);
-                }
-                // SAFETY: `in_ptr + blen..in_ptr + blen + tail_len`
-                // is valid region for reads and `tail_len` is smaller than `BS`.
-                // we have verified that `blen + bs <= out_len`, in other words,
-                // `out_ptr + blen..out_ptr + blen + bs` is valid region
-                // for writes.
-                let out_block = unsafe {
-                    core::ptr::copy_nonoverlapping(
-                        self.in_ptr.add(blen),
-                        tail_in.as_mut_ptr(),
-                        tail_len,
-                    );
-                    &mut *(self.out_ptr.add(blen) as *mut Array<u8, BS>)
-                };
-                P::pad(&mut tail_in, tail_len);
-                Some(out_block)
-            }
+
+        let Some(tail_block) = tail_block else {
+            let tail_inout = None;
+            return Ok(PaddedInOutBuf { blocks, tail_inout });
         };
-        Ok(PaddedInOutBuf {
-            blocks,
-            tail_in,
-            tail_out,
-        })
+
+        let blocks_byte_len = blocks_len * bs;
+        let reserve_len = out_len - blocks_byte_len;
+        if reserve_len < tail_block.len() {
+            return Err(PadError);
+        }
+        // SAFETY: we checked that the out buffer has enough bytes in reserve
+        let tail_out: &mut Array<u8, BS> = unsafe {
+            let tail_out_ptr = out_ptr.add(blocks_byte_len);
+            &mut *(tail_out_ptr.cast())
+        };
+
+        let tail_inout = Some((tail_block, tail_out));
+
+        Ok(PaddedInOutBuf { blocks, tail_inout })
     }
 }
 
 /// Variant of [`InOutBuf`] with optional padded tail block.
 #[cfg(feature = "block-padding")]
+#[allow(clippy::type_complexity)]
 pub struct PaddedInOutBuf<'inp, 'out, BS: ArraySize> {
     blocks: InOutBuf<'inp, 'out, Array<u8, BS>>,
-    tail_in: Array<u8, BS>,
-    tail_out: Option<&'out mut Array<u8, BS>>,
+    tail_inout: Option<(Array<u8, BS>, &'out mut Array<u8, BS>)>,
 }
 
 #[cfg(feature = "block-padding")]
@@ -221,20 +215,20 @@ impl<'out, BS: ArraySize> PaddedInOutBuf<'_, 'out, BS> {
 
     /// Get padded tail block.
     ///
-    /// For paddings with `P::TYPE = PadType::Reversible` it always returns `Some`.
+    /// Most padding implementations always return `Some`.
     #[inline(always)]
-    #[allow(clippy::needless_option_as_deref)]
     pub fn get_tail_block(&mut self) -> Option<InOut<'_, '_, Array<u8, BS>>> {
-        match self.tail_out.as_deref_mut() {
-            Some(out_block) => Some((&self.tail_in, out_block).into()),
-            None => None,
-        }
+        self.tail_inout.as_mut().map(|(in_block, out_block)| {
+            let in_block = &*in_block;
+            let out_block = &mut **out_block;
+            InOut::from((in_block, out_block))
+        })
     }
 
     /// Convert buffer into output slice.
     #[inline(always)]
     pub fn into_out(self) -> &'out [u8] {
-        let total_blocks = if self.tail_out.is_some() {
+        let total_blocks = if self.tail_inout.is_some() {
             self.blocks.len() + 1
         } else {
             self.blocks.len()