+++ block-padding/CHANGELOG.md
@@ -5,17 +5,23 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## 0.4.0 (unreleased)
+### Added
+- `Padding::pad_detached` method ([#1225])
+
 ### Changed
 - Migrated from `generic-array` to `hybrid-array` ([#944])
 - Edition changed to 2024 and MSRV bumped to 1.85 ([#1149])
 - Merged `RawPadding` and `Padding` traits ([#1217])
+- Renamed `UnpadError` to `Error` ([#1225])
 
 ### Removed
 - `Block` type alias ([#1217])
+- `PadType` enum and associated type on the `Padding` trait ([#1225])
 
 [#944]: https://github.com/RustCrypto/utils/pull/944
 [#1149]: https://github.com/RustCrypto/utils/pull/1149
 [#1217]: https://github.com/RustCrypto/utils/pull/1217
+[#1225]: https://github.com/RustCrypto/utils/pull/1225
 
 ## 0.3.3 (2023-04-02)
 ### Added